apps/pdf_worker/requirements.txt
pikepdf==9.4.0
pdfminer.six==20240706
functions-framework==3.*
